requirements.txt
# Loopy-Munch Integration Dependencies
# Production requirements for the integration service

# Web framework
Flask==2.3.3
Flask-CORS==4.0.0

# HTTP requests
requests==2.31.0

# JSON and data handling
python-dotenv==1.0.0

# Date and time utilities
python-dateutil==2.8.2

# Logging and utilities (included in Python standard library)
# logging - built-in
# datetime - built-in
# json - built-in
# typing - built-in
# os - built-in

# Development dependencies (optional)
# pytest==7.4.3
# black==23.12.1
# flake8==6.1.0

gunicorn==21.2.0
Werkzeug==3.0.1
cryptography==41.0.7
PyJWT==2.8.0 